requirements.txt
numpy
pandas
matplotlib
umap-learn